CMS-658: Split up access context and provider
