+++ frontend/src/contexts/AccessContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+// Create a new context for access control,
+// to be used in the AccessProvider and ProtectedRoute components
+export default createContext();
